+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [[*next-version*]] - YYYY-MM-DD
+### Added
+- Added separate WP page for managing services.
+
+### Removed
+- Removed the bookings metabox in the new/edit download page.
 
 ## [0.1-alpha22] - 2018-10-08
 ### Added